docs: add content to getting started page (#598)
These changes enhance the "Getting Started" documentation in the website.

Add basic getting started guide which includes the following:
  - Installing via npm, pnpm or yarn
  - importing the module and displaying first graph
  - Explaining core graph parts (vertex, edge and group).
  - How to handle events (click and cells moved event)
  - Basic vertex styling


The documentation and the README of the repository contain some duplication and the related duplicated parts are marked to ease the future maintenance.
The new contents will be reworked once the mxGraph tutorial has been integrated into the documentation.