fix: popup menu story caused has more submenus open (#572)
Add a workaround to remove unreleased/destroyed PopupMenu and MaxWindow when navigating from one story to the next. Otherwise, the window or menu remains displayed in the next story.
The reason for this is the failure to clean up the graph and destroy its resources correctly when dismantling the story. Stories are defined using Storybook's HTML support, which does not provide an unmount/destroy function.

The solution is to delete DOM elements with an mxPopupMenu or mxWindow class.
This has no impact on functionality, as graph initialization within the story will add them later.

---------

Co-authored-by: Thomas Bouffard <27200110+tbouffard@users.noreply.github.com>