fix(TooltipHandler): initialize the div when tooltip is shown (#591)
Previously, the TooltipHandler code didn't work in the same way as the mxGraph code. We created the div in the constructor, not when the `show` function was called.
This created a div even though the tooltip was never shown.

To solve the problem, we restore mxGraph's logic. Initialization of the DOM element has been moved from the constructor to the `init` function, which is called if the `show` function is called and the DOM element has not been initialized.