fix(story): correctly set the perimeter on labels at page load (#589)
This is a workaround for the `PerimeterOnLabelBoundsP story. It's the Storybook rendering that's causing this problem, nothing in the base code.

Before the workaround, the `getPerimeterPoint` function was called before anything was painted on the canvas. So, when calling `node.getBBox()` to get the smallest bounded rectangle for the SVGElement, we would get the value 0 because nothing had been painted at the time the code was executed.

The solution is to use `setTimeout` for `batchUpdate`, so that it's updated after the container has been loaded into history and everything has been painted.