feat!: make edge bend points easier to manage (#633)
Previously in `mxGraph`, this was done by updating the properties on the prototype of the `EdgeHandler` class.
This didn't work anymore in `maxGraph`.
It is now possible to configure them back by using new properties available in `EdgeHandlerConfig`

BREAKING CHANGES: the following properties have been removed from `EdgeHandler` and are replaced by global  configuration in `EdgeHandlerConfig`:
  - addEnabled --> addBendOnShiftClickEnabled
  - removeEnabled --> removeBendOnShiftClickEnabled
  - virtualBendOpacity
  - virtualBendsEnabled