fix: remove some circular dependencies (#663)
The problems were detected by rollup application in the `maxgraph-integration-examples` repository.

Introduce the usage of the `madge` tool to detect such problems, especially because we already add a similar in the past fixed in df98075.
The detection of the circular dependencies now runs as part of the CI build, so similar problems will be detected earlier than in the past.